pyproject.toml
[project]
name = "find-screenshots-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.0",
    "rich>=14.3.1",
    "typer>=0.21.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
